Remake super list for colors + starting tilesets

+++ RPG Paper Maker/Engine/SuperListBox.cs
@@ -14,10 +14,14 @@ namespace RPG_Paper_Maker
 {
     public partial class SuperListBox : UserControl
     {
-        public List<SuperListItem> ModelList;
+        public List<SuperListItem> ModelList { get { return listBox.Items.Cast<SuperListItem>().ToList(); } }
         public Type DialogKind;
+        public Type TypeItem;
         public ListBox[] ListBoxes;
-        public int Max;
+        public SuperListItem CopiedItem = null;
+        public int Min, Max;
+        public System.Timers.Timer DragTimer = new System.Timers.Timer(20);
+        public bool CanDrag = false;
 
         public SuperListBox()
         {
@@ -28,18 +32,21 @@ namespace RPG_Paper_Maker
         // InitializeListParameters
         // -------------------------------------------------------------------
 
-        public void InitializeListParameters(ListBox[] list, List<SuperListItem> model, Type type, int max)
+        public void InitializeListParameters(ListBox[] list, List<SuperListItem> modelList, Type type, Type typeItem, int min, int max)
         {
             ListBoxes = list;
-            ModelList = model;
             DialogKind = type;
+            TypeItem = typeItem;
+            Min = min;
             Max = max;
 
-            for (int i = 0; i < ModelList.Count; i++)
+            listBox.FormattingEnabled = false;
+            DragTimer.Elapsed += new System.Timers.ElapsedEventHandler(DoDrag);
+
+            for (int i = 0; i < modelList.Count; i++)
             {
-                listBox.Items.Add(WANOK.GetStringList((i + 1), ModelList[i].Name));
+                listBox.Items.Add(modelList[i]);
             }
-            if (ModelList.Count < WANOK.MAX_COLORS) listBox.Items.Add(WANOK.ListBeginning);
         }
 
         // -------------------------------------------------------------------
@@ -69,49 +76,43 @@ namespace RPG_Paper_Maker
 
         public void EditItem()
         {
-            SuperListDialog dialog = listBox.SelectedIndex < ModelList.Count ? 
-                (SuperListDialog)Activator.CreateInstance(DialogKind, ModelList[listBox.SelectedIndex]) : 
-                (SuperListDialog)Activator.CreateInstance(DialogKind, 
-                            BindingFlags.CreateInstance |
-                            BindingFlags.Public |
-                            BindingFlags.Instance |
-                            BindingFlags.OptionalParamBinding, null, new object[] { Type.Missing }, CultureInfo.CurrentCulture);
-            if (dialog.ShowDialog() == DialogResult.OK)
+            // If double clic is opening a window...
+            if (DialogKind != null)
             {
-                int index = listBox.SelectedIndex;
-                if (index >= ModelList.Count && ModelList.Count < Max) listBox.Items.Add(WANOK.ListBeginning);
-                if (index >= ModelList.Count) ModelList.Add(dialog.GetObject());
-                else ModelList[index] = dialog.GetObject();
-                listBox.Items.RemoveAt(index);
-                listBox.Items.Insert(index, WANOK.GetStringList((index + 1), dialog.GetObject().Name));
+                SuperListDialog dialog = (SuperListDialog)Activator.CreateInstance(DialogKind, listBox.Items[listBox.SelectedIndex]);
+
+                if (dialog.ShowDialog() == DialogResult.OK)
+                {
+                    int index = listBox.SelectedIndex;
+                    listBox.Items.RemoveAt(index);
+                    listBox.Items.Insert(index, dialog.GetObject());
+                }
+            }
+            // If the settings are directly on the right panel...
+            else
+            {
+
             }
         }
 
         // -------------------------------------------------------------------
-        // AddItem
+        // CopyItem
         // -------------------------------------------------------------------
 
-        public void AddItem()
+        public void CopyItem()
         {
-            if (ModelList.Count == Max) MessageBox.Show("Maximum reached.", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-            else {
-                SuperListDialog dialog = (SuperListDialog)Activator.CreateInstance(DialogKind,
-                            BindingFlags.CreateInstance |
-                            BindingFlags.Public |
-                            BindingFlags.Instance |
-                            BindingFlags.OptionalParamBinding, null, new object[] { Type.Missing }, CultureInfo.CurrentCulture);
+            CopiedItem = ((SuperListItem)listBox.Items[listBox.SelectedIndex]).CreateCopy();
+        }
 
-                if (dialog.ShowDialog() == DialogResult.OK)
-                {
-                    listBox.Items.Insert(listBox.SelectedIndex + 1, WANOK.GetStringList((listBox.SelectedIndex + 2), dialog.GetObject().Name));
-                    ModelList.Insert(listBox.SelectedIndex + 1, dialog.GetObject());
-                    if (ModelList.Count == WANOK.MAX_COLORS) listBox.Items.RemoveAt(listBox.Items.Count - 1);
-                    for (int i = listBox.SelectedIndex + 2; i < ModelList.Count; i++)
-                    {
-                        listBox.Items[i] = WANOK.GetStringList(i+1, ModelList[i].Name);
-                    }
-                }
-            }
+        // -------------------------------------------------------------------
+        // PasteItem
+        // -------------------------------------------------------------------
+
+        public void PasteItem()
+        {
+            CopiedItem.Id = ((SuperListItem)listBox.Items[listBox.SelectedIndex]).Id;
+            listBox.Items[listBox.SelectedIndex] = CopiedItem;
+            CopiedItem = CopiedItem.CreateCopy();
         }
 
         // -------------------------------------------------------------------
@@ -120,17 +121,10 @@ namespace RPG_Paper_Maker
 
         public void DeleteItem()
         {
-            if (listBox.Items.Count == 2) MessageBox.Show("You need at least one element.", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-            else {
-                int index = listBox.SelectedIndex;
-                listBox.Items.RemoveAt(index);
-                ModelList.RemoveAt(index);
-                for (int i = index; i < ModelList.Count; i++)
-                {
-                    listBox.Items[i] = WANOK.GetStringList(i + 1, ModelList[i].Name);
-                }
-                listBox.SelectedIndex = index;
-            }
+            int id = ((SuperListItem)listBox.Items[listBox.SelectedIndex]).Id;
+            SuperListItem defaultValue = (SuperListItem)Activator.CreateInstance(TypeItem);
+            defaultValue.Id = id;
+            listBox.Items[listBox.SelectedIndex] = defaultValue;
         }
 
         // -------------------------------------------------------------------
@@ -139,6 +133,14 @@ namespace RPG_Paper_Maker
 
         private void listBox_MouseDown(object sender, MouseEventArgs e)
         {
+            // If left clic, can drag and drop
+            if (e.Button == MouseButtons.Left)
+            {
+                if (listBox.SelectedItem == null) return;
+                if (!DragTimer.Enabled) DragTimer.Start();
+            }
+
+            // If right clic, open ContextMenu
             if (e.Button == MouseButtons.Right)
             {
                 int index = listBox.IndexFromPoint(e.X, e.Y);
@@ -146,10 +148,7 @@ namespace RPG_Paper_Maker
                 listBox.SelectedIndex = index;
                 if (listBox.SelectedIndex != -1)
                 {
-                    bool exists = listBox.SelectedIndex < ModelList.Count;
-                    ItemEdit.Text = exists ? "Edit" : "New element";
-                    ItemAdd.Enabled = exists;
-                    ItemDelete.Enabled = exists;
+                    ItemPaste.Enabled = CopiedItem != null;
                     contextMenuStrip.Show(listBox, e.Location);
                 }
             }
@@ -177,12 +176,21 @@ namespace RPG_Paper_Maker
         }
 
         // -------------------------------------------------------------------
-        // ItemAdd_Click
+        // ItemCopy_Click
         // -------------------------------------------------------------------
 
-        private void ItemAdd_Click(object sender, EventArgs e)
+        private void ItemCopy_Click(object sender, EventArgs e)
         {
-            AddItem();
+            CopyItem();
+        }
+
+        // -------------------------------------------------------------------
+        // ItemPaste_Click
+        // -------------------------------------------------------------------
+
+        private void ItemPaste_Click(object sender, EventArgs e)
+        {
+            PasteItem();
         }
 
         // -------------------------------------------------------------------
@@ -200,9 +208,84 @@ namespace RPG_Paper_Maker
 
         private void listBox_KeyDown(object sender, KeyEventArgs e)
         {
-            if (listBox.SelectedIndex != -1 && listBox.SelectedIndex < listBox.Items.Count - 1 && e.KeyCode == Keys.Delete)
+            if (listBox.SelectedIndex != -1)
             {
-                DeleteItem();
+                if (e.KeyCode == Keys.Delete) DeleteItem();
+                if (e.Control && e.KeyCode == Keys.C) CopyItem();
+                if (e.Control && e.KeyCode == Keys.V && CopiedItem != null) PasteItem();
+            }
+        }
+
+        // -------------------------------------------------------------------
+        // DoDrag
+        // -------------------------------------------------------------------
+
+        private void DoDrag(object sender, System.Timers.ElapsedEventArgs e)
+        {
+            CanDrag = true;
+            DragTimer.Stop();
+        }
+
+        // -------------------------------------------------------------------
+        // listBox_DragDrop
+        // -------------------------------------------------------------------
+
+        private void listBox_DragDrop(object sender, DragEventArgs e)
+        {
+            Point point = listBox.PointToClient(new Point(e.X, e.Y));
+            int newIndex = listBox.IndexFromPoint(point);
+            if (newIndex < 0) newIndex = listBox.Items.Count - 1;
+            object data = e.Data.GetData(TypeItem);
+
+            listBox.Items.Remove(data);
+            listBox.Items.Insert(newIndex, data);
+
+            /*
+            if (newIndex < beforeIndex)
+            {
+                for (int i = 0; i < OrderList.Count; i++) if (OrderList[i] >= newIndex && OrderList[i] < beforeIndex) OrderList[i]++;
+            }
+            else if (newIndex > beforeIndex)
+            {
+                for (int i = 0; i < OrderList.Count; i++) if (OrderList[i] > beforeIndex && OrderList[i] <= newIndex) OrderList[i]--;
+            }
+            OrderList[idItemMoved - 1] = newIndex;
+            
+            
+            for (int i = 0; i < OrderList.Count; i++)
+            {
+                if (((SuperListItem)listBox.Items[OrderList[i]]).Id != i + 1) throw new Exception();
+            }*/
+        }
+
+        // -------------------------------------------------------------------
+        // listBox_DragOver
+        // -------------------------------------------------------------------
+
+        private void listBox_DragOver(object sender, DragEventArgs e)
+        {
+            e.Effect = DragDropEffects.Move;
+        }
+
+        // -------------------------------------------------------------------
+        // listBox_MouseUp
+        // -------------------------------------------------------------------
+
+        private void listBox_MouseUp(object sender, MouseEventArgs e)
+        {
+            CanDrag = false;
+        }
+
+        // -------------------------------------------------------------------
+        // listBox_MouseMove
+        // -------------------------------------------------------------------
+
+        private void listBox_MouseMove(object sender, MouseEventArgs e)
+        {
+            if (CanDrag)
+            {
+                listBox.DoDragDrop(listBox.SelectedItem, DragDropEffects.Move);
+                CanDrag = false;
             }
         }
 
@@ -213,6 +296,34 @@ namespace RPG_Paper_Maker
         private void listBox_MouseEnter(object sender, EventArgs e)
         {
             listBox.Focus();
+        }
+
+        // -------------------------------------------------------------------
+        // button_Click
+        // -------------------------------------------------------------------
+
+        private void button_Click(object sender, EventArgs e)
+        {
+            DialogEnterNumber dialog = new DialogEnterNumber(listBox.Items.Count, Min, Max, ModelList);
+            if (dialog.ShowDialog() == DialogResult.OK)
+            {
+                // Suppress
+                if (listBox.Items.Count > dialog.Value)
+                {
+                    int nb = listBox.Items.Count - dialog.Value;
+                    for (int i = 0; i < nb; i++) listBox.Items.RemoveAt(dialog.Value);
+                }
+                // Add
+                else if (listBox.Items.Count < dialog.Value)
+                {
+                    int nb = dialog.Value - listBox.Items.Count;
+                    for (int i = 0; i < nb; i++) {
+                        SuperListItem defaultValue = (SuperListItem)Activator.CreateInstance(TypeItem);
+                        defaultValue.Id = listBox.Items.Count + 1;
+                        listBox.Items.Add(defaultValue);
+                    }
+                }
+            }
         }
     }
 }
